move train_optimization.py to train
